+++ crates/runner-host/crates/web-server/src/lib.rs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 use std::net::SocketAddr;
+use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
@@ -22,6 +23,7 @@ use tokio::sync::{mpsc, oneshot};
 use warp::ws::{Message, WebSocket};
 use warp::Filter;
 use wasefire_logger as log;
+use wasefire_protocol::applet::ExitStatus;
 use web_common::{ButtonState, Command, Component};
 
 #[derive(Debug, Copy, Clone)]
@@ -35,11 +37,11 @@ pub struct Client {
 }
 
 impl Client {
-    pub async fn new(addr: SocketAddr, events: mpsc::Sender<Event>) -> Result<Self> {
+    pub async fn new(dir: PathBuf, addr: SocketAddr, events: mpsc::Sender<Event>) -> Result<Self> {
         let (sender, mut receiver) = oneshot::channel();
         let client = Arc::new(Mutex::new(Some((sender, events))));
 
-        let static_files = warp::fs::dir("crates/web-client/dist");
+        let static_files = warp::fs::dir(dir);
         let ws = warp::path("board")
             .and(warp::ws())
             .and(warp::any().map(move || client.clone()))
@@ -71,6 +73,14 @@ impl Client {
 
     pub fn set_led(&self, state: bool) {
         self.send(Command::Set { component_id: LED_ID, state });
+    }
+
+    pub fn start(&self) {
+        self.send(Command::Start);
+    }
+
+    pub fn exit(&self, status: ExitStatus) {
+        self.send(Command::Exit { status });
     }
 
     fn send(&self, command: Command) {