+++ crates/cli-tools/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### Minor
 
+- Add `action::PlatformInfo` to print platform serial and version
+- Add `cmd::spawn()` for more control on command execution
+- Add `fs::remove_dir_all()` to remove a directory recursively
+- Add `fs::rename()` to rename a file
 - Add `cargo` and `changelog` modules and features
 - Handle more errors during platform discovery and `action::PlatformReboot`
 - Extend `fs::write()` first parameter to set the `OpenOptions` too
@@ -35,4 +39,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 7 -->
+<!-- Increment to skip CHANGELOG.md test: 8 -->